started working on node model and node model manager

+++ src/sdk/message/message_type_gen.js
@@ -247,7 +247,7 @@ void ${typename}::update_static_header() {
   }`;
     data += `
   static_headers.message_size = message_size;
-  static_headers.header_size = header_size;
+  static_headers.header_size = (uint16_t)header_size;
 }
 
 }  // namespace dsa
